requirements.txt
flask
flask_bootstrap4
flask_sqlalchemy
